requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
attrs==25.4.0
dotenv==0.9.9
frozenlist==1.8.0
gunicorn==23.0.0
idna==3.11
multidict==6.7.0
packaging==25.0
pillow==12.0.0
propcache==0.4.1
python-dotenv==1.2.1
redis==7.1.0
uvloop==0.22.1
yarl==1.22.0
